requirements.txt
Flask
Flask-Mail
psycopg2-binary
Werkzeug
config
flask_cors 
openai
azure-ai-inference
azure-core
python-dotenv  
